requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.2
django-filter>=23.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
celery>=5.3
redis>=5.0

# NLP and AI dependencies (optional - for enhanced AI features)
# Uncomment these when ready to use external AI services:
# openai>=1.0
# anthropic>=0.3
# tiktoken>=0.5
# numpy>=1.24
# scipy>=1.11
